+++ src/main.cpp
@@ -23,7 +23,7 @@ int main(int argc, char const *argv[])
     layer->loadInput(data, 4, 4);
     layer->calculateOutput();
     
-    int w, h;
+    unsigned long w, h;
     double *res = layer->getOutput(w, h);
     assert(w == 2);
     assert(h == 2);
